Pruebas de aceptación publicadas por la cátedra exitosas!

+++ workspace/HyperMarket/src/ar/uba/fi/tecnicasdedisenio/grupo8/hypermarket/promocion/DescuentoVenta.java
@@ -13,7 +13,7 @@ public class DescuentoVenta {
 	private IVenta venta;
 	private RepositorioPromociones repositorioPromociones;
 	private Collection<ItemDescuento> itemsDescuento;
-	private int importeDescuentoVenta;
+	private double importeDescuentoVenta;
 	private boolean importeDescuentoVentaCalculado; 
 
 	public DescuentoVenta(IVenta venta,RepositorioPromociones repositorioPromociones) {
